reafactor: seperate GradientContainer
