+++ role_dice/lib/main.dart
@@ -9,32 +9,3 @@ void main() {
     debugShowCheckedModeBanner: false,
   ));
 }
-
-class GradientContainer extends StatelessWidget {
-  const GradientContainer({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: const BoxDecoration(
-        gradient: LinearGradient(
-          colors: <Color>[
-            Color.fromARGB(255, 18, 8, 111),
-            Color.fromARGB(255, 31, 27, 150),
-          ],
-          begin: Alignment.bottomRight,
-          end: Alignment.topLeft,
-        ),
-      ),
-      child: const Center(
-        child: Text(
-          '#100DayChallenge',
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: 28,
-          ),
-        ),
-      ),
-    );
-  }
-}
